# Replace aom-av1 with svt-av1 for saving webm videos, use preset 6 + yuv420p10le pixel format