This PR ~adds support for saving .webm outputs using codec svt-av1 and adds an options for using yuv420p10le pixel format~ switches aom-av1 codec for svt-av1 + yuv420p10le pixel format.

svt-av1 is significantly faster at encoding compared to libaom-av1 and yuv420p10le provides higher quality av1 output without size increase.

Note: av added support for libsvtav1 in v14.1.0, the requirements.txt are upped to reflect this.

# Comparison
## before (5d0d4ee9)
Took **407s** to save an 81 frame wan webm using libaom-av1, crf 10.

1.8MB, VMAF 96.9885 vs an animated png output.

## pr code with svt-av1 & yuv420p10le selected
Took **1s** with libsvtav1 (preset 6, crf 10).

2.3MB, VMAF 97.23115 _(note pixel format provides better quality vs aom at yuv420)_

### preset time & quality
I've set preset 6 for svt-av1, would be straightforward to make this configurable (perhaps later).
* preset 1: 19.46s, VMAF 97.43995
* preset 3: 3.88s, VMAF 97.39643
* preset 6: 0.95s, VMAF 97.23115
* preset 8: 0.67s, VMAF 96.918396

### vs webp
Result is ~2.3MB vs ~6MB .webp using slowest/quality:90, webp also having visibly lower quality.

# Implementation notes
~I tried to make this non-breaking so any existing usage of av1 will continue to use aom-av1 + yuv420. However, I would  recommend always using svt-av1 and yuv420p10le for any CPU-encoding scenario (which seem best for short videos like this as they are better quality than gpu encoding). So perhaps the aom-av1 option should be phased out or somehow discouraged in future. Maybe renamed "aom-av1 (slow)" or something?~