@comfyanonymous 2025-04-22
You can just switch the "av1" to use svt, I don't think many people are currently using it because of how slow it is. As long as quality is the same or better it's good.