@alexheretic 2025-04-22
> You can just switch the "av1" to use svt, I don't think many people are currently using it because of how slow it is. As long as quality is the same or better it's good.

Technically aom-av1 produces a little better quality, but it's _so_ much slower that it isn't ever worth it for mortals. But with the pixel format switch we do actually get both 400x faster and better quality (though very slightly larger files, still smaller than the alternatives).

Ok so you'd prefer I keep the interface unchanged and just switch out for svt-av1 + yuv420p10le under the hood? 

This might break people how git pull without pip-installing, as they'll perhaps have outdated av. Is that ok? Or should we put in a warning?