@alexheretic 2025-04-22
I've updated the approach following your suggestion.