@comfyanonymous 2025-04-22
14.1.0 is older than the time the dependency was introduced in ComfyUI so that part is fine. Just to make sure do you know if yuv420p10le compatible with every av1 decoder?