@alexheretic 2025-04-22
>  Just to make sure do you know if yuv420p10le compatible with every av1 decoder?

I don't 100% know. I've never had an issue with it, I think it's part of the av1 spec but I'm not 100% sure. So it's possible that some decoders somewhere could have issue with it. I can add it back as optional if you prefer?